fozzie-components@v2.16.0 – Making `sassOptions` vue config reusable
